+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ar.io/wayfinder-core
 
+## 1.0.1-alpha.0
+
+### Patch Changes
+
+- aa5700e: Improve telemetry configuration for browsers and chrome extensions
+
 ## 1.0.0
 
 ### Major Changes