+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ar.io/wayfinder-react
 
+## 1.0.2-alpha.0
+
+### Patch Changes
+
+- Updated dependencies [aa5700e]
+  - @ar.io/wayfinder-core@1.0.1-alpha.0
+
 ## 1.0.1
 
 ### Patch Changes